+++ src/SecondQuantizedAlgebra.jl
@@ -4,8 +4,9 @@ using SymbolicUtils: SymbolicUtils, simplify, substitute, add_worker
 using Symbolics: Symbolics, Num, expand, @variables, build_function, symbolic_to_float
 using TermInterface: TermInterface
 
-import QuantumInterface: ⊗, tensor, expect, basis, dagger
+import QuantumInterface: expect, basis, dagger
 using QuantumInterface: AbstractOperator, StateVector, Basis
+import TensorCore: ⊗, tensor
 
 using Combinatorics: with_replacement_combinations
 using Latexify: Latexify, latexify, @latexrecipe